+++ templates/login.html
@@ -1,30 +1,120 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <meta charset="UTF-8" />
+    <meta charset="UTF-8">
     <title>Login - KuberaVault</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/login.css') }}">
+    <style>
+        html, body {
+            height: 100%;
+            margin: 0;
+            padding: 0;
+        }
+        body {
+            min-height: 100vh;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            {% if login_bg %}
+            background: url("{{ url_for('static', filename='images/' ~ login_bg) }}") no-repeat center center fixed;
+            background-size: cover;
+            background-color: #e5e5e5;
+            {% else %}
+            background: #e5e5e5;
+            {% endif %}
+        }
+        .container {
+            background: rgba(255,255,255,0.92);
+            max-width: 400px;
+            width: 100%;
+            margin: 0 auto;
+            padding: 2.5rem 2rem 2rem 2rem;
+            border-radius: 14px;
+            box-shadow: 0 4px 32px rgba(0,0,0,0.18);
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+        }
+        h2 {
+            margin-top: 0;
+            margin-bottom: 1.5rem;
+            font-family: 'Merriweather', serif;
+            color: #2a1f1e;
+        }
+        form {
+            width: 100%;
+            display: flex;
+            flex-direction: column;
+            gap: 1rem;
+        }
+        input[type="text"], input[type="password"] {
+            width: 100%;
+            padding: 0.75rem;
+            border: 1px solid #ccc;
+            border-radius: 6px;
+            font-size: 1rem;
+        }
+        button[type="submit"] {
+            padding: 0.75rem;
+            background: #7c4d03;
+            color: #fff;
+            border: none;
+            border-radius: 6px;
+            font-size: 1rem;
+            cursor: pointer;
+            transition: background 0.2s;
+        }
+        button[type="submit"]:hover {
+            background: #a67c52;
+        }
+        .messages {
+            width: 100%;
+            margin: 1rem 0 0.5rem 0;
+        }
+        .messages ul {
+            padding-left: 1.2em;
+            margin: 0;
+        }
+        .msg {
+            font-size: 0.98em;
+        }
+        .msg.success { color: #2e7d32; }
+        .msg.error { color: #c62828; }
+        p {
+            margin-top: 1.5rem;
+            font-size: 0.98em;
+        }
+        a {
+            color: #7c4d03;
+            text-decoration: underline;
+        }
+    </style>
 </head>
 <body>
+    {% if 'user_id' in session %}
+    <div style="position: absolute; top: 1rem; right: 1rem;">
+        <a href="{{ url_for('logout') }}" class="btn btn-danger" style="padding: 0.5em 1em; border-radius: 6px; background: #c62828; color: #fff; text-decoration: none;">
+            <i class="fas fa-sign-out-alt" aria-label="Logout"></i> Logout
+        </a>
+    </div>
+    {% endif %}
     <div class="container">
         <h2>Login to KuberaVault</h2>
-
-        {% with messages = get_flashed_messages(with_categories=true) %}
-          {% if messages %}
-            <div class="messages">
-              {% for category, message in messages %}
-                <p class="msg {{ category }}">{{ message }}</p>
-              {% endfor %}
-            </div>
-          {% endif %}
-        {% endwith %}
-
         <form method="POST" action="{{ url_for('login') }}">
             <input type="text" name="username" placeholder="Username" required />
             <input type="password" name="password" placeholder="Password" required />
             <button type="submit">Login</button>
         </form>
-
+        {% with messages = get_flashed_messages(with_categories=true) %}
+          {% if messages %}
+            <div class="messages">
+              <ul>
+              {% for category, message in messages %}
+                <li class="msg {{ category }}">{{ message }}</li>
+              {% endfor %}
+              </ul>
+            </div>
+          {% endif %}
+        {% endwith %}
         <p>Don't have an account? <a href="{{ url_for('register') }}">Register here</a></p>
     </div>
 </body>
