+++ templates/profile.html
@@ -21,35 +21,68 @@
                 <div class="header-nav">
                     <ul class="nav-links">
                         <li class="nav-item">
-                            <a href="/dashboard" class="nav-link active" title="Dashboard">
-                                <i class="fas fa-home"></i>
+                            <a href="/dashboard" class="nav-link active" title="Dashboard" aria-label="Go to Dashboard">
+                                <i class="fas fa-home" aria-label="Dashboard"></i>
+                                <span>Dashboard</span>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="/transactions" class="nav-link" title="Transactions">
-                                <i class="fas fa-exchange-alt"></i>
+                            <a href="/transactions" class="nav-link" title="Transactions" aria-label="Go to Transactions">
+                                <i class="fas fa-exchange-alt" aria-label="Transactions"></i>
+                                <span>Transactions</span>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="/financial-learning" class="nav-link" title="Financial Learning">
-                                <i class="fas fa-book-reader"></i>
+                            <a href="/budget" class="nav-link" title="Budget" aria-label="Go to Budget">
+                                <i class="fas fa-chart-pie" aria-label="Budget"></i>
+                                <span>Budget</span>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="/budget" class="nav-link" title="Budget">
-                                <i class="fas fa-chart-pie"></i>
+                            <a href="/financial-learning" class="nav-link" title="Financial Learning" aria-label="Go to Financial Learning">
+                                <i class="fas fa-book-reader" aria-label="Financial Learning"></i>
+                                <span>Financial Learning</span>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="/faq" class="nav-link" title="FAQ">
-                                <i class="fas fa-question-circle"></i>
+                            <a href="/scholarships-loans" class="nav-link" title="Scholarships & Loans" aria-label="Go to Scholarships and Loans">
+                                <i class="fas fa-graduation-cap" aria-label="Scholarships & Loans"></i>
+                                <span>Scholarships & Loans</span>
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a href="/profile" class="nav-link active" title="Profile">
-                                <i class="fas fa-user-circle"></i>
+                            <a href="/faq" class="nav-link" title="FAQ" aria-label="Go to FAQ">
+                                <i class="fas fa-question-circle" aria-label="FAQ"></i>
+                                <span>FAQ</span>
                             </a>
                         </li>
+                        {% if 'user_id' not in session %}
+                        <li class="nav-item">
+                            <a href="{{ url_for('login') }}" class="nav-link" title="Login" aria-label="Go to Login">
+                                <i class="fas fa-sign-in-alt" aria-label="Login"></i>
+                                <span>Login</span>
+                            </a>
+                        </li>
+                        <li class="nav-item">
+                            <a href="{{ url_for('register') }}" class="nav-link" title="Register" aria-label="Go to Register">
+                                <i class="fas fa-user-plus" aria-label="Register"></i>
+                                <span>Register</span>
+                            </a>
+                        </li>
+                        {% else %}
+                        <li class="nav-item">
+                            <a href="{{ url_for('profile') }}" class="nav-link" title="Profile" aria-label="Go to Profile">
+                                <i class="fas fa-user-circle" aria-label="Profile"></i>
+                                <span>Profile</span>
+                            </a>
+                        </li>
+                        <li class="nav-item">
+                            <a href="{{ url_for('logout') }}" class="nav-link" title="Logout" aria-label="Logout">
+                                <i class="fas fa-sign-out-alt" aria-label="Logout"></i>
+                                <span>Logout</span>
+                            </a>
+                        </li>
+                        {% endif %}
                     </ul>
                 </div>
             </div>
